+++ packages/@expo/config-plugins/src/ios/__tests__/BuildScheme-test.ts
@@ -13,6 +13,8 @@ const fsReal = jest.requireActual('fs') as typeof fs;
 jest.mock('fs');
 
 describe(getRunnableSchemesFromXcodeproj, () => {
+  const platform = 'ios';
+
   beforeAll(async () => {
     vol.fromJSON(
       {
@@ -29,7 +31,7 @@ describe(getRunnableSchemesFromXcodeproj, () => {
     vol.reset();
   });
   it(`parses for runnable schemes`, async () => {
-    const schemes = getRunnableSchemesFromXcodeproj('/app');
+    const schemes = getRunnableSchemesFromXcodeproj('/app', platform);
     expect(schemes).toStrictEqual([
       { name: 'multitarget', osType: 'iOS', type: 'com.apple.product-type.application' },
       { name: 'shareextension', osType: 'iOS', type: 'com.apple.product-type.app-extension' },
@@ -38,6 +40,8 @@ describe(getRunnableSchemesFromXcodeproj, () => {
 });
 
 describe(getApplicationTargetNameForSchemeAsync, () => {
+  const platform = 'ios';
+
   describe('single build action entry', () => {
     beforeAll(async () => {
       vol.fromJSON(
@@ -54,13 +58,13 @@ describe(getApplicationTargetNameForSchemeAsync, () => {
     });
 
     it('returns the target name for existing scheme', async () => {
-      const target = await getApplicationTargetNameForSchemeAsync('/app', 'testproject');
+      const target = await getApplicationTargetNameForSchemeAsync('/app', platform, 'testproject');
       expect(target).toBe('testproject');
     });
 
     it('throws if the scheme does not exist', async () => {
       await expect(() =>
-        getApplicationTargetNameForSchemeAsync('/app', 'nonexistentscheme')
+        getApplicationTargetNameForSchemeAsync('/app', platform, 'nonexistentscheme')
       ).rejects.toThrow(/does not exist/);
     });
   });
@@ -80,19 +84,21 @@ describe(getApplicationTargetNameForSchemeAsync, () => {
     });
 
     it('returns the target name for existing scheme', async () => {
-      const target = await getApplicationTargetNameForSchemeAsync('/app', 'testproject');
+      const target = await getApplicationTargetNameForSchemeAsync('/app', platform, 'testproject');
       expect(target).toBe('testproject');
     });
 
     it('throws if the scheme does not exist', async () => {
       await expect(() =>
-        getApplicationTargetNameForSchemeAsync('/app', 'nonexistentscheme')
+        getApplicationTargetNameForSchemeAsync('/app', platform, 'nonexistentscheme')
       ).rejects.toThrow(/does not exist/);
     });
   });
 });
 
 describe(getArchiveBuildConfigurationForSchemeAsync, () => {
+  const platform = 'ios';
+
   beforeAll(async () => {
     vol.fromJSON(
       {
@@ -110,6 +116,7 @@ describe(getArchiveBuildConfigurationForSchemeAsync, () => {
   it('returns build configuration name for existing scheme', async () => {
     const buildConfiguration = await getArchiveBuildConfigurationForSchemeAsync(
       '/app',
+      platform,
       'testproject'
     );
     expect(buildConfiguration).toBe('Release');
@@ -117,7 +124,7 @@ describe(getArchiveBuildConfigurationForSchemeAsync, () => {
 
   it('throws if the scheme does not exist', async () => {
     await expect(() =>
-      getArchiveBuildConfigurationForSchemeAsync('/app', 'nonexistentscheme')
+      getArchiveBuildConfigurationForSchemeAsync('/app', platform, 'nonexistentscheme')
     ).rejects.toThrow(/does not exist/);
   });
 });