+++ packages/@expo/config-plugins/src/ios/utils/Xcodeproj.ts
@@ -22,6 +22,7 @@ import xcode, {
 import pbxFile from 'xcode/lib/pbxFile';
 
 import { trimQuotes } from './string';
+import { ModPlatform } from '../../Plugin.types';
 import { addWarningIOS } from '../../utils/warnings';
 import * as Paths from '../Paths';
 
@@ -37,17 +38,18 @@ export type ConfigurationListEntry = [string, XCConfigurationList];
 
 export type ConfigurationSectionEntry = [string, XCBuildConfiguration];
 
-export function getProjectName(projectRoot: string) {
-  const sourceRoot = Paths.getSourceRoot(projectRoot);
+export function getProjectName(projectRoot: string, platform: ModPlatform) {
+  const sourceRoot = Paths.getSourceRoot(projectRoot, platform);
   return path.basename(sourceRoot);
 }
 
 export function resolvePathOrProject(
-  projectRootOrProject: string | XcodeProject
+  projectRootOrProject: string | XcodeProject,
+  platform: ModPlatform
 ): XcodeProject | null {
   if (typeof projectRootOrProject === 'string') {
     try {
-      return getPbxproj(projectRootOrProject);
+      return getPbxproj(projectRootOrProject, platform);
     } catch {
       return null;
     }
@@ -72,10 +74,14 @@ function sanitizedNameForProjects(name: string) {
 // the ios project paths. Overall this function needs to be revamped, just a
 // placeholder for now! Make this more robust when we support applying config
 // at any time (currently it's only applied on eject).
-export function getHackyProjectName(projectRoot: string, config: ExpoConfig): string {
+export function getHackyProjectName(
+  projectRoot: string,
+  platform: ModPlatform,
+  config: ExpoConfig
+): string {
   // Attempt to get the current ios folder name (apply).
   try {
-    return getProjectName(projectRoot);
+    return getProjectName(projectRoot, platform);
   } catch {
     // If no iOS project exists then create a new one (eject).
     const projectName = config.name;
@@ -332,8 +338,8 @@ export function ensureGroupRecursively(project: XcodeProject, filepath: string):
 /**
  * Get the pbxproj for the given path
  */
-export function getPbxproj(projectRoot: string): XcodeProject {
-  const projectPath = Paths.getPBXProjectPath(projectRoot);
+export function getPbxproj(projectRoot: string, platform: ModPlatform): XcodeProject {
+  const projectPath = Paths.getPBXProjectPath(projectRoot, platform);
   const project = xcode.project(projectPath);
   project.parseSync();
   return project;